Move class and change namespace.
