+++ IsraelHiking.API/Controllers/SearchController.cs
@@ -7,7 +7,7 @@ using NetTopologySuite.Features;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using IsraelHiking.API.Services.Poi;
+using IsraelHiking.API.Services;
 
 namespace IsraelHiking.API.Controllers;
