+++ Tests/IsraelHiking.API.Tests/Controllers/PointsOfInterestControllerTests.cs
@@ -1,7 +1,6 @@
 ﻿using IsraelHiking.API.Controllers;
 using IsraelHiking.API.Executors;
 using IsraelHiking.API.Services;
-using IsraelHiking.API.Services.Poi;
 using IsraelHiking.Common;
 using IsraelHiking.Common.Api;
 using IsraelHiking.Common.Configuration;
